+++ member-app-external-api/build.gradle
@@ -32,6 +32,11 @@ jar {
 }
 
 dependencies {
+    implementation 'javax.xml.bind:jaxb-api:2.3.0'
+    implementation 'io.jsonwebtoken:jjwt:0.9.1'
+    implementation 'com.google.code.gson:gson:2.9.1'
+    implementation 'org.modelmapper:modelmapper:3.1.0'
+    implementation 'org.springframework.data:spring-data-jpa'
     implementation 'org.springframework.boot:spring-boot-starter-web'
     testImplementation 'org.projectlombok:lombok:1.18.26'
     compileOnly 'org.projectlombok:lombok'
