+++ member-app-external-api/src/main/java/io/howstheairtoday/memberappexternalapi/security/config/RootConfig.java
@@ -1,4 +1,4 @@
-package io.howstheairtoday.memberappexternalapi.config;
+package io.howstheairtoday.memberappexternalapi.security.config;
 
 import org.modelmapper.ModelMapper;
 import org.modelmapper.convention.MatchingStrategies;
