+++ member-app-external-api/src/test/java/io/howstheairtoday/memberappexternalapi/security/util/JWTUtilTest.java
@@ -1,4 +1,4 @@
-package io.howstheairtoday.memberappexternalapi.util;
+package io.howstheairtoday.memberappexternalapi.security.util;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
@@ -11,6 +11,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
+import io.howstheairtoday.memberappexternalapi.security.util.JWTUtil;
 import io.jsonwebtoken.MalformedJwtException;
 import lombok.extern.log4j.Log4j2;
