+++ module-core/build.gradle
@@ -23,14 +23,8 @@ jar {
 dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-web'
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
-
     api 'org.springframework.boot:spring-boot-starter-validation'
-    api 'com.google.code.gson:gson:2.9.1'
-    api 'org.springframework.data:spring-data-jpa'
-    api 'org.modelmapper:modelmapper:3.1.0'
     api 'org.springframework.boot:spring-boot-starter-security'
-    api 'javax.xml.bind:jaxb-api:2.3.0'
-    api 'io.jsonwebtoken:jjwt:0.9.1'
     testImplementation 'org.springframework.security:spring-security-test'
 }
