--- module-core/src/main/java/io/howstheairtoday/modulecore/ModuleCoreApplication.java
@@ -1,12 +1,0 @@
-// package io.howstheairtoday.modulecore;
-//
-// import org.springframework.boot.SpringApplication;
-// import org.springframework.boot.autoconfigure.SpringBootApplication;
-//
-// @SpringBootApplication
-// public class ModuleCoreApplication {
-//
-//     public static void main(String[] args) {
-//         SpringApplication.run(ModuleCoreApplication.class, args);
-//     }
-// }
